+++ en/docusaurus.config.ts
@@ -143,7 +143,7 @@ const config: Config = {
           items: [
             { label: 'Overview', to: '/docs/get-started/overview-and-architecture' },
             { label: 'Install', to: '/docs/get-started/install' },
-            { label: 'Quick starts', to: '/docs/get-started/quick-start-automation' },
+            { label: 'Quick starts', to: '/docs/get-started/build-automation' },
           ],
         },
         {